Move chapter components into new dir
